==> scripts/01_data_retrieve_wrangle.Rmd <==
---
title: "population data: retrieval and wrangling"
subtitle: "Census 2021: Census Subdivisions"
author: "Martin Monkman"
date: "2023-06-29"
output: html_document
---

## setup

```{r setup}
# packages
library(tidyverse)
library(readxl)
# packages: utilities
library(here)
library(glue)

# cansim
# reference page: https://mountainmath.github.io/cansim/index.html
library(cansim) 


# create list of Vancouver Island CDs numbers
vi_cd_list <- c("17", "19", "21", "23", "24", "26", "27", "43", "45")

five_year_bins <- c(
  "0 to 4 years",
  "5 to 9 years",
  "10 to 14 years",
  "15 to 19 years",
  "20 to 24 years",
  "25 to 29 years",
  "30 to 34 years",
  "35 to 39 years",
  "40 to 44 years",
  "45 to 49 years",
  "50 to 54 years",
  "55 to 59 years",
  "60 to 64 years",
  "65 to 69 years",
  "70 to 74 years",
  "75 to 79 years",
  "80 to 84 years",
  "85 to 89 years",
  "90 to 94 years",
  "95 to 99 years",
  "100 years and over"  
)

five_year_bins_90 <- c(
  "0 to 4 years",
  "5 to 9 years",
  "10 to 14 years",
  "15 to 19 years",
  "20 to 24 years",
  "25 to 29 years",
  "30 to 34 years",
  "35 to 39 years",
  "40 to 44 years",
  "45 to 49 years",
  "50 to 54 years",
  "55 to 59 years",
  "60 to 64 years",
  "65 to 69 years",
  "70 to 74 years",
  "75 to 79 years",
  "80 to 84 years",
  "85 to 89 years",
  "90 years and older"  
)

```


## Census 2021 - single year


Age (in single years), average age and median age and gender: Census metropolitan areas, census agglomerations and census subdivisions1
Frequency: Occasional

Table: 98-10-0021-01

Release date: 2022-04-27

Geography: Census subdivision, Census metropolitan area, Census agglomeration

Universe: Total Population, including institutional residents, 2021 and 2016 censuses — 100% data

Variable List: Age (in single years), average age and median age (128), Census year (2), Gender (3a)


How to cite: Statistics Canada. Table 98-10-0021-01  Age (in single years), average age and median age and gender: Census metropolitan areas, census agglomerations and census subdivisions
DOI: https://doi.org/10.25318/9810002101-eng


https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810002101

_NOTE: 98-10-0021-01 does NOT include Tofino and Ucluelet, perhaps due to boundary changes between the Census years_



Age (in single years), average age and median age and gender: Canada, provinces and territories, census divisions, census subdivisions and dissemination areas
Frequency: Occasional

Table: 98-10-0023-01

Release date: 2022-04-27

Geography: Canada, Province or territory, Census division, Census subdivision, Dissemination area

Universe: Total population, including institutional residents, 2021 Census — 100% data

Variable List: Age (in single years), average age and median age (128), Gender (3)

https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810002301

How to cite: Statistics Canada. Table 98-10-0023-01  Age (in single years), average age and median age and gender: Canada, provinces and territories, census divisions, census subdivisions and dissemination areas
DOI: https://doi.org/10.25318/9810002301-eng


```{r}
df_pop_csd_detail <- read_csv(here::here("data_statcan", "98100023.csv")) |> 
  janitor::clean_names()

write_rds(df_pop_csd_detail, here::here("data_statcan", "98100023.rds"))
```


```{r}
df_pop_csd_detail <- read_rds(here::here("data_statcan", "98100023.rds"))
df_pop_csd_detail
```

### CSD - BC

```{r}
df_pop_csd_bc <- df_pop_csd_detail |> 
  filter(str_detect(dguid, "2021A000559"))

#### ---- UNNECESSARY WITH THIS FILE
#|> 
## strip geo into component parts
#  # geocode with parenthesis
#  mutate(geo_code = str_match(geo, "\\(\\d{7}\\)")) |> 
#  mutate(geo_code = str_remove(geo_code, "^\\(")) |> 
#  mutate(geo_code = str_remove(geo_code, "\\)$")) |> 
#  # cd and csd numbers separately
#  mutate(geo_cd = as.character(str_sub(dguid, start = 12, end = 13)))  |> 
#  mutate(geo_csd = as.character(str_sub(dguid, start = 14, end = 16)))  |> 
#  # place name
#  mutate(geo_name = str_match(geo, "^.*\\(\\d")) |> 
#  mutate(geo_name = str_remove(geo_name, " \\(\\d$")) |> 
#  # geo type
#  mutate(geo_type = str_match(geo, " \\D+$")) |> 
#  mutate(geo_type = str_trim(geo_type))


# separate CD and CSD values
df_pop_csd_bc <- df_pop_csd_bc |> 
  # geocode with parenthesis
  mutate(geo_cd = as.character(str_sub(dguid, start = 12, end = 13)))  |> 
  mutate(geo_csd = as.character(str_sub(dguid, start = 14, end = 16))) 


# separate CD and CSD values
df_pop_csd_bc <- df_pop_csd_bc |> 
  # geocode with parenthesis
  mutate(geo_cd = as.character(str_sub(dguid, start = 12, end = 13)))  |> 
  mutate(geo_csd = as.character(str_sub(dguid, start = 14, end = 16))) 
  

#write_rds(df_pop_csd_bc, "data_statcan/df_pop_csd_bc.rds")
```

```{r}

#df_pop_csd_bc <- read_rds("data_statcan/df_pop_csd_bc.rds")
  
df_pop_csd_bc_5yr <- df_pop_csd_bc |> 
  filter(age_in_single_years_average_age_and_median_age_128 %in%
           five_year_bins) |> 
  rename(
    age_grp = age_in_single_years_average_age_and_median_age_128,
    pop_total = gender_3_total_gender_1,
    pop_men = gender_3_men_2,
    pop_women = gender_3_women_3
) |> 
  mutate(age_grp = as_factor(age_grp))

levels(df_pop_csd_bc_5yr$age_grp)

#write_rds(df_pop_csd_bc_5yr, "data_statcan/df_pop_csd_bc_5yr.rds")

```


### BC 5 year

```{r read_bc_5yr}

#df_pop_csd_bc_5yr <- read_rds("data_statcan/df_pop_csd_bc_5yr.rds")

```


NOTE: "geo_name" ("geo" in original file) is _not_ unique! Use dguid as unique identifier

```{r}

# Vancouver Island CDs

df_pop_csd_vi_5yr <- df_pop_csd_bc_5yr |> 
  select(
    geo_name = geo,
    dguid,
    geo_cd,
    geo_csd,
    age_grp,
    pop_total,
    pop_men,
    pop_women
  ) |> 
  filter(
    geo_cd %in% vi_cd_list
    )

df_pop_csd_vi_5yr

# check

df_pop_csd_vi_5yr |> 
  summarise(
    total_pop = sum(pop_total, na.rm = TRUE)
  )

```

```{r}

# calculate gender as pct of total, by dguid
df_pop_csd_vi_5yr <- df_pop_csd_vi_5yr |> 
  group_by(dguid) |> 
  mutate(
    total_pop_csd = sum(pop_total, na.rm = TRUE),
    total_men_csd = sum(pop_men, na.rm = TRUE),
    total_women_csd = sum(pop_women, na.rm = TRUE)
    ) 

  # checks -- doesn't add up due to random rounding
#  mutate(
#    total_total = total_men_csd + total_women_csd
#  ) |> 
#  mutate(
#    total_check = total_pop_csd - total_total
#  ) |> 
#  select(geo, total_check)
#

df_pop_csd_vi_5yr <- df_pop_csd_vi_5yr |> 
  mutate(
    pct_men = pop_men / total_pop_csd * 100,
    pct_women = pop_women / total_pop_csd * 100
  ) |> 
  mutate(pct_age_grp = pct_men + pct_women) |> 
  ungroup()

# check
df_pop_csd_vi_5yr |> 
  group_by(geo_name) |> 
  summarise(total_pct_100 = sum(pct_age_grp))

df_pop_csd_vi_5yr |> 
  distinct(geo_name, geo_cd, geo_csd)

df_pop_csd_vi_5yr

write_rds(df_pop_csd_vi_5yr, "data_statcan/df_pop_csd_vi_5yr.rds")
write_csv(df_pop_csd_vi_5yr, "data_statcan/df_pop_csd_vi_5yr.csv")


```


### CD - BC

```{r}
df_pop_cd_comox <- df_pop_csd_detail |> 
  filter(str_detect(dguid, "2021A00035926"))

df_pop_cd_comox

```

```{r}
df_pop_cd_comox_5yr <- df_pop_cd_comox |> 
  filter(age_in_single_years_average_age_and_median_age_128 %in%
           five_year_bins) |> 
  rename(
    age_grp = age_in_single_years_average_age_and_median_age_128,
    pop_total = gender_3_total_gender_1,
    pop_men = gender_3_men_2,
    pop_women = gender_3_women_3
) |> 
  mutate(age_grp = as_factor(age_grp))

levels(df_pop_cd_comox_5yr$age_grp)

#write_rds(df_pop_cd_comox_5yr, here::here("data_statcan", "df_pop_cd_comox_5yr.rds"))

#df_pop_cd_comox_5yr <- read_rds(here::here("data_statcan", "df_pop_cd_comox_5yr.rds"))

```



## Economic Region

Note: for other regional boundaries, see
https://www150.statcan.gc.ca/n1/pub/91-214-x/91-214-x2023001-eng.htm


[Annual demographic estimates, economic regions: Interactive dashboard](https://www150.statcan.gc.ca/n1/pub/71-607-x/71-607-x2020020-eng.htm)


**CANSIM TABLE**

Population estimates, July 1, by economic region, 2016 boundaries1, 2, 3, 4, 5, 6
Frequency: Annual

Table: 17-10-0137-01

Release date: 2023-01-11

Geography: Canada, Province or territory, Economic region


https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1710013701




```{r}

df_pop_er_canada <- get_cansim("17-10-0137-01")

df_pop_er_canada <- df_pop_er_canada |> 
  janitor::clean_names()

#write_rds(df_pop_er_canada, here::here("data_statcan", "df_pop_er_canada.rds"))
#df_pop_er_canada <- read_rds(here::here("data_statcan", "df_pop_er_canada.rds"))


unique(df_pop_er_canada$geo)

unique(df_pop_er_canada$age_group)

```


### Vancouver Island & Coast


#### 5 year bins

```{r}

df_pop_er_vic <- df_pop_er_canada |> 
  filter(geo == "Vancouver Island and Coast, British Columbia") |> 
  filter(age_group %in% five_year_bins_90)

df_pop_er_vic

# median age
df_pop_er_canada |> 
  filter(geo == "Vancouver Island and Coast, British Columbia") |> 
  filter(age_group == "Median age")

```
